Create PathfinderConfigurationBuilder
Creating Pathfinder configurations isn't fun, as the core code behind them is rather outdated and I don't have the energy to refactor it as of now. To fix this problem, or better put, put a band-aid over it, the configuration builder class allows you to create a configuration with as little as two parameters - a drivetrain and an odometry system.